Extract user package
