+++ src/main/java/com/code4ro/legalconsultation/common/security/CurrentUser.java
@@ -1,7 +1,7 @@
 package com.code4ro.legalconsultation.common.security;
 
 import com.code4ro.legalconsultation.model.persistence.ApplicationUser;
-import com.code4ro.legalconsultation.model.persistence.UserRole;
+import com.code4ro.legalconsultation.user.model.persistence.UserRole;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Getter;
 import org.springframework.security.core.GrantedAuthority;
