+++ src/main/java/com/code4ro/legalconsultation/controller/DocumentController.java
@@ -6,6 +6,7 @@ import com.code4ro.legalconsultation.model.persistence.DocumentConsolidated;
 import com.code4ro.legalconsultation.model.persistence.DocumentExportFormat;
 import com.code4ro.legalconsultation.model.persistence.DocumentMetadata;
 import com.code4ro.legalconsultation.service.api.DocumentService;
+import com.code4ro.legalconsultation.user.model.dto.UserDto;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import lombok.RequiredArgsConstructor;
