+++ src/main/java/com/code4ro/legalconsultation/converters/DocumentConsolidatedMapper.java
@@ -2,6 +2,7 @@ package com.code4ro.legalconsultation.converters;
 
 import com.code4ro.legalconsultation.model.dto.DocumentConsolidatedDto;
 import com.code4ro.legalconsultation.model.persistence.DocumentConsolidated;
+import com.code4ro.legalconsultation.user.mapper.UserMapper;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring",
