+++ src/main/java/com/code4ro/legalconsultation/model/dto/DocumentConsolidatedDto.java
@@ -1,6 +1,7 @@
 package com.code4ro.legalconsultation.model.dto;
 
 import com.code4ro.legalconsultation.model.dto.documentnode.DocumentNodeDto;
+import com.code4ro.legalconsultation.user.model.dto.UserDto;
 import lombok.Getter;
 import lombok.Setter;
