+++ src/main/java/com/code4ro/legalconsultation/model/persistence/ApplicationUser.java
@@ -1,5 +1,6 @@
 package com.code4ro.legalconsultation.model.persistence;
 
+import com.code4ro.legalconsultation.user.model.persistence.User;
 import lombok.Getter;
 import lombok.Setter;
