+++ src/main/java/com/code4ro/legalconsultation/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.code4ro.legalconsultation.repository;
 
-import com.code4ro.legalconsultation.model.persistence.User;
+import com.code4ro.legalconsultation.user.model.persistence.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
