+++ src/main/java/com/code4ro/legalconsultation/service/api/InvitationService.java
@@ -2,7 +2,7 @@ package com.code4ro.legalconsultation.service.api;
 
 import com.code4ro.legalconsultation.model.dto.SignUpRequest;
 import com.code4ro.legalconsultation.model.persistence.Invitation;
-import com.code4ro.legalconsultation.model.persistence.User;
+import com.code4ro.legalconsultation.user.model.persistence.User;
 
 public interface InvitationService {
     Invitation getInvitation(String code);
