+++ src/main/java/com/code4ro/legalconsultation/service/api/MailApi.java
@@ -2,7 +2,7 @@ package com.code4ro.legalconsultation.service.api;
 
 import com.code4ro.legalconsultation.model.persistence.Invitation;
 import com.code4ro.legalconsultation.model.persistence.DocumentMetadata;
-import com.code4ro.legalconsultation.model.persistence.User;
+import com.code4ro.legalconsultation.user.model.persistence.User;
 import java.util.List;
 
 public interface MailApi {
