+++ src/main/java/com/code4ro/legalconsultation/service/impl/ApplicationUserService.java
@@ -4,10 +4,10 @@ import com.code4ro.legalconsultation.common.exceptions.I18nError;
 import com.code4ro.legalconsultation.common.exceptions.LegalValidationException;
 import com.code4ro.legalconsultation.model.dto.SignUpRequest;
 import com.code4ro.legalconsultation.model.persistence.ApplicationUser;
-import com.code4ro.legalconsultation.model.persistence.User;
-import com.code4ro.legalconsultation.model.persistence.UserRole;
+import com.code4ro.legalconsultation.user.model.persistence.User;
 import com.code4ro.legalconsultation.repository.ApplicationUserRepository;
 import com.code4ro.legalconsultation.service.api.InvitationService;
+import com.code4ro.legalconsultation.user.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.CachePut;
 import org.springframework.http.HttpStatus;
