+++ src/main/java/com/code4ro/legalconsultation/service/impl/DocumentServiceImpl.java
@@ -2,10 +2,13 @@ package com.code4ro.legalconsultation.service.impl;
 
 import com.code4ro.legalconsultation.converters.DocumentConsolidatedMapper;
 import com.code4ro.legalconsultation.converters.PdfHandleMapper;
-import com.code4ro.legalconsultation.converters.UserMapper;
+import com.code4ro.legalconsultation.user.mapper.UserMapper;
 import com.code4ro.legalconsultation.model.dto.*;
 import com.code4ro.legalconsultation.model.persistence.*;
 import com.code4ro.legalconsultation.service.api.*;
+import com.code4ro.legalconsultation.user.model.dto.UserDto;
+import com.code4ro.legalconsultation.user.model.persistence.User;
+import com.code4ro.legalconsultation.user.service.UserService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import com.code4ro.legalconsultation.service.api.DocumentNodeService;
