+++ src/main/java/com/code4ro/legalconsultation/service/impl/InvitationServiceImpl.java
@@ -4,7 +4,7 @@ import com.code4ro.legalconsultation.common.exceptions.LegalValidationException;
 import com.code4ro.legalconsultation.model.dto.SignUpRequest;
 import com.code4ro.legalconsultation.model.persistence.Invitation;
 import com.code4ro.legalconsultation.model.persistence.InvitationStatus;
-import com.code4ro.legalconsultation.model.persistence.User;
+import com.code4ro.legalconsultation.user.model.persistence.User;
 import com.code4ro.legalconsultation.repository.InvitationRepository;
 import com.code4ro.legalconsultation.service.api.InvitationService;
 import org.apache.commons.lang3.RandomStringUtils;
