+++ src/main/java/com/code4ro/legalconsultation/user/model/persistence/UserRole.java
@@ -1,4 +1,4 @@
-package com.code4ro.legalconsultation.model.persistence;
+package com.code4ro.legalconsultation.user.model.persistence;
 
 public enum UserRole {
     CONTRIBUTOR,
