+++ src/test/java/com/code4ro/legalconsultation/controller/AuthControllerIntegrationTest.java
@@ -7,7 +7,7 @@ import com.code4ro.legalconsultation.model.dto.LoginRequest;
 import com.code4ro.legalconsultation.model.dto.SignUpRequest;
 import com.code4ro.legalconsultation.model.persistence.Invitation;
 import com.code4ro.legalconsultation.model.persistence.InvitationStatus;
-import com.code4ro.legalconsultation.model.persistence.User;
+import com.code4ro.legalconsultation.user.model.persistence.User;
 import com.code4ro.legalconsultation.repository.ApplicationUserRepository;
 import com.code4ro.legalconsultation.repository.InvitationRepository;
 import com.code4ro.legalconsultation.repository.UserRepository;
