+++ src/test/java/com/code4ro/legalconsultation/controller/InvitationControllerTest.java
@@ -4,7 +4,7 @@ import com.code4ro.legalconsultation.common.controller.AbstractControllerIntegra
 import com.code4ro.legalconsultation.factory.RandomObjectFiller;
 import com.code4ro.legalconsultation.model.persistence.Invitation;
 import com.code4ro.legalconsultation.model.persistence.InvitationStatus;
-import com.code4ro.legalconsultation.model.persistence.User;
+import com.code4ro.legalconsultation.user.model.persistence.User;
 import com.code4ro.legalconsultation.repository.InvitationRepository;
 import com.code4ro.legalconsultation.repository.UserRepository;
 import org.junit.Test;
