+++ src/test/java/com/code4ro/legalconsultation/service/ApplicationUserServiceTest.java
@@ -4,11 +4,11 @@ import com.code4ro.legalconsultation.common.exceptions.LegalValidationException;
 import com.code4ro.legalconsultation.factory.RandomObjectFiller;
 import com.code4ro.legalconsultation.model.dto.SignUpRequest;
 import com.code4ro.legalconsultation.model.persistence.ApplicationUser;
-import com.code4ro.legalconsultation.model.persistence.User;
+import com.code4ro.legalconsultation.user.model.persistence.User;
 import com.code4ro.legalconsultation.repository.ApplicationUserRepository;
 import com.code4ro.legalconsultation.service.api.InvitationService;
 import com.code4ro.legalconsultation.service.impl.ApplicationUserService;
-import com.code4ro.legalconsultation.service.impl.UserService;
+import com.code4ro.legalconsultation.user.service.UserService;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
