+++ src/test/java/com/code4ro/legalconsultation/service/DocumentServiceTest.java
@@ -5,7 +5,7 @@ import com.code4ro.legalconsultation.factory.RandomObjectFiller;
 import com.code4ro.legalconsultation.model.dto.DocumentConsolidatedDto;
 import com.code4ro.legalconsultation.model.dto.DocumentMetadataDto;
 import com.code4ro.legalconsultation.model.persistence.DocumentConsolidated;
-import com.code4ro.legalconsultation.model.persistence.User;
+import com.code4ro.legalconsultation.user.model.persistence.User;
 import com.code4ro.legalconsultation.service.api.MailApi;
 import com.code4ro.legalconsultation.model.persistence.*;
 import com.code4ro.legalconsultation.service.api.CommentService;
@@ -14,7 +14,7 @@ import com.code4ro.legalconsultation.service.export.PDFExporter;
 import com.code4ro.legalconsultation.service.impl.DocumentConsolidatedService;
 import com.code4ro.legalconsultation.service.impl.DocumentMetadataService;
 import com.code4ro.legalconsultation.service.impl.DocumentServiceImpl;
-import com.code4ro.legalconsultation.service.impl.UserService;
+import com.code4ro.legalconsultation.user.service.UserService;
 import com.code4ro.legalconsultation.service.impl.pdf.PDFServiceImpl;
 import org.junit.Test;
 import org.junit.runner.RunWith;
