+++ src/test/java/com/code4ro/legalconsultation/service/InvitationServiceTest.java
@@ -5,7 +5,7 @@ import com.code4ro.legalconsultation.factory.RandomObjectFiller;
 import com.code4ro.legalconsultation.model.dto.SignUpRequest;
 import com.code4ro.legalconsultation.model.persistence.Invitation;
 import com.code4ro.legalconsultation.model.persistence.InvitationStatus;
-import com.code4ro.legalconsultation.model.persistence.User;
+import com.code4ro.legalconsultation.user.model.persistence.User;
 import com.code4ro.legalconsultation.repository.InvitationRepository;
 import com.code4ro.legalconsultation.service.impl.InvitationServiceImpl;
 import org.junit.Test;
